requirements.txt
boto3
Flask
psycopg2-binary
nltk
scikit-learn
